requirements.txt
argon2-cffi>=23.1.0
cryptography>=43.0.0
fido2>=1.1.3
